+++ test/utest-online/testlinkapi_online_test.py
@@ -252,6 +252,9 @@ class TestLinkAPIOnlineTestCase(unittest.TestCase):
             self.client.uploadAttachment(attachemantFile, 4712, 'nodes_hierarchy',
                         title='title 4713', description='descr. 4714')
 
+    def test_testLinkVersion(self):
+        response = self.client.testLinkVersion()
+        self.assertRegexpMatches(response, '\d*\.\d*\.\d*')
         
 if __name__ == "__main__":
     #import sys;sys.argv = ['', 'Test.testName']